cardmaker.ai/requirements.txt
Pillow==10.3.0
requests==2.32.3
psycopg2-binary==2.9.9
